Clean up experimental work
This branch was intended to be very exploratory - both in terms of
finding reasonable solutions to making Rust and C# work together (like
kInvoke) and in terms of testing the performance of Rust vs C# (like
everything imported from LDJ50 on the Frontend side, and the deacon
rules on the AI side).

Now that we have our answers - Rust is ~3600 times faster, and it's
totally feasible to build a C#/Rust bridge - we can delete a lot of
stuff.