Implement terminal_value
Since this depends on the number of moves available to the given player,
this commit also changes SpiceState to generate the full list of moves
on SpiceState creation, so we can simply use that list's len(). This
also lets us change the MoveIterator to a vec::IntoIter over SpiceMoves,
and simplifies the move generation code a little.