Enable code coverage reports during tests
Note that we depend on com.unity.testtools.codecoverage v1.1.1 instead
of the latest (v1.2.3 as of this commit). This is because of a bug
introduced in v1.2.2 which prevents generic code from being recorded -
the package authors claim that the bug is fixed in v1.2.3, but I still
consistently reproduced the bug with that version. There is also a bug
with Unity's built-in code coverage API, which hasn't been resolved.
When both bugs have been fixed, we should update to latest. See:

https://forum.unity.com/threads/how-to-use-code-coverage-with-c-generics.1337024/
https://issuetracker.unity3d.com/issues/coverage-package-fails-to-gather-any-coverage-for-generic-methods
https://issuetracker.unity3d.com/issues/coverage-dot-getstatsfor-api-shows-0-hits-for-generic-methods